[refactor] Remove indirection around ReleaseTags
There is no need for this to be in two separate classes, one will do
